+++ autogpt/core/runner/__init__.py
@@ -1,0 +1,4 @@
+import autogpt.core.status
+
+status = autogpt.core.status.Status.IN_PROGRESS
+handover_notes = "A basic client and server is being sketched out. It is at the idea stage."
